wrap paper to TRA
I think the paper is ok for the full paper TRA.
There is a lot of room for improvement, as marked in comments, when submitting to https://www.sciencedirect.com/journal/computers-environment-and-urban-systems/about/call-for-papers#environmentally-and-socially-sustainable-mobility in late november.
@Robinlovelace do you want to take a look at the output, before the submission next week?

Still need to HEAT to be in good heath to make sure that the SE estimates remain the same after the changes.